requirements.txt
flask
flask-cors
deepface
pillow
numpy
textblob
gtts
speechrecognition
google-generativeai
pydub
